fix: automations trigger wrong output path
I have noticed that when using a Node-RED trigger in automations it always triggered the wrong (second / index 1) output path of my switch. While browsing through the code, as I thought, this is an issue that must be related to my wrongdoings, as I couldn't be the first one noticing this, I stumbled upon that the `output_path` is parsed as a boolean only for automations. Which explains why in most of my tests output 1 was chosen, as most strings evaluate to a boolish "true". This change fixed parsing the output path as string instead.